requirements.txt
openmetadata-ingestion==1.8.0
pandas
importlib-metadata
pyarrow
boto3
pyyaml